rename insee:seed command to migrate:insee
